Restructure EulerAcousticsCouplingCallback
The callback no longer constructs the pure flow ODE itself, instead
it has to be constructed by the user and passed to the constructor.
That way the `Union` shenanigans can be avoided as the
Euler integrator can be constructed inside of the constructor of
the callback. This leads to a significant reduction of allocations.
Furthermore, the user can choose which algorithm is used to solve
the pure flow problem and can provide kwargs to the pure flow solver
such as callbacks (e.g. for saving restart files).