Make tspan a non-keyword argument
AveragingCallback currently expects the averaging time span to be passed
as a keyword argument (tspan). Usually the variable name tspan is also used
for controlling time integration. Therefore, it is not possible to control
both variables separately via trixi_include so AveragingCallback no
longer uses this keyword argument to avoid this.