Add MPI support for non-uniform meshes (2D TreeMesh) (#898)
* Add MPI mortar container

Adds a new container `MPIL2MortarContainer2D` which implements the same
functionality as `L2MortarContainer2D` but for MPI mortars, i.e. mortars
where the neighboring elements do not live on a single MPI rank.

* Implement solver functions for MPI mortars

* Implement MPI mortar connectivity and communication routines

* Add tests for MPI mortars

* Fix MPI AMR

The AMR implementation only reinitialized containers when there is
something to refine/coarsen/rebalance. The initialization of `MPICache`
uses all-to-all communication, therefore MPI stalled if only a subset
of all MPI ranks refined/coarsened/rebalanced something.

* Apply suggestions from code review

Co-authored-by: Michael Schlottke-Lakemper <michael@sloede.com>

* Put `mpi_isparallel()` before code blocks relevant for parallel execution

* Exit counting routines early if MPI is not used

* Add comment about how `large_sides` is determined for MPI mortars

* Use auxiliary variable to simplify index calculation

* Rename pos -> position

* Use regular if-conditions

* first -> Base.first

A variable called `first` is defined earlier in this function, which
causes some issues. Therefore `Base.first` needs to be called including
the module name.

* Add comments to tests where serial/parallel errors are the same

* Exit init! functions early if container size is 0

* Use iseven in init_mortars

* Determine MPI neighbor ranks using vcat exclusively instead of reduce

* Remove obsolete mpi_isparallel() checks

* Use any instead of !all, switch to 2-arg version of all

* Switch order of if and for loops in prolong2mpimortars

* Change signature of `element_solutions_to_mortars!`

This function now expects a mortar container instead of the complete
cache as the first argument. That way, the function can be used for
both the regular mortars and the MPI mortars in parallel settings.

Co-authored-by: Hendrik Ranocha <ranocha@users.noreply.github.com>
Co-authored-by: Michael Schlottke-Lakemper <michael@sloede.com>