Add MPI support for conforming `P4estMesh`es in 2D (#977)
* Check if p4est supports MPI

If the p4est library does not support MPI, it is not allowed to
pass an MPI communicator when creating a new p4est object. If it does
support MPI, an MPI communicator has to be passed even if only one
process is used.

* Distinguish between serial and parallel P4estMeshes

This enables adding methods specialized on the parallel type for MPI runs.

* Add containers for parallel P4estMesh

Adds a new container for the interfaces at the MPI domain boundaries
and methods for initializing such container and adds specialized
versions for some initialization methods for the existing containers
for parallel runs (e.g. to distinguish between regular and mpi interfaces).

* Add parallel solver functions

* Minor cleanup, hacky fix for unstructed BCs

* WIP mesh, solution and restart files for parallel P4estMeshes

* fix mesh I/O

* unsafe_load. -> unsafe_wrap

* Finish parallel I/O for `P4estMesh`

* Add parallel p4est tests

* rename p4est_new_ghost_layer -> new_ghost_layer_p4est

* Rename containers_2d_parallel.jl -> containers_parallel_2d.jl

* Create unified `ncells` function for `P4estMesh`

* Switch order of `RealT` and `IsParallel` params in `P4estMesh`

This avoids a few unnecessary "where {IsParallel}" and makes the code
more readable overall.

* TODO comments

* Use P4est.jl v0.3.0

- New P4est.jl version supports MPI, hence enable `ParallelP4estMesh` tests
- `p4est_has_mpi()` function no longer required as it's not built into
  P4est.jl

Co-authored-by: Michael Schlottke-Lakemper <michael@sloede.com>
Co-authored-by: Hendrik Ranocha <ranocha@users.noreply.github.com>