Add support for non-conforming `ParallelP4estMesh`es in 2D (#1080)
* Add support for non-conforming `ParallelP4estMesh`es in 2D

Similar to the existing TreeMesh-code, an additional container for MPI
mortars, i.e. mortars whose adjacent elements are distributed across
several processes, was added. This container holds lists containing
the IDs and corresponding positions of locally available elements adjacent
to the mortar. Since the P4estMesh is curved, the normal directions on the
element surfaces are needed to calculate the mortar fluxes. For the serial
mortars, the normal directions from the small elements are used. In the
parallel case, these may not be available, hence the normal directions
are calculated once during container initialization and then distributed
to the other involved processes.

Conceptually, the solver functions for calculating the MPI mortar fluxes
work the same way as their serial counterparts.

The functions related to MPI communication and connectivity work the same
as their TreeMesh counterparts. In particular, NaNs are used as
sentinel values to determine if data has been received for a particular
position of a mortar.

In addition, the handling of p4est ghost layers was adjusted.
Previously it was only constructed during initialization of the MPI
communication and connectivity data structures and destroyed
afterwards. Now the P4estMesh has an additional field for the ghost layer
and it is automatically constructed when creating the mesh. In addition,
the ghost layer gets updated when the containers are (re-)initialized.
This change was necessary because hanging ghost sides are not always
correctly identified as such during `iterate_p4est` when no ghost layer
is passed and thus the ghost layer is necessary for all `iterate_p4est`
calls when creating the cache.

* Fix container (re-)initialization

* Implement parallel AMR including load balancing for 2D P4estMesh

* Update src/meshes/p4est_mesh.jl

Co-authored-by: Michael Schlottke-Lakemper <michael@sloede.com>

* Add explanatory comments in `init_surfaces_iter_face_inner` function

* n_small_elems -> n_small_elements

* mpi -> MPI in comments

* Use true/false instead of 1/0 when comparing `is_hanging` or `is_ghost`

* Clarify what `buffer_mortar_indices` does

* Add comment about NaN conversion

Co-authored-by: Michael Schlottke-Lakemper <michael@sloede.com>

* Change `buffer_mortar_indices` comment

* Use `copy` to copy `global_first_quadrant` during load balancing

Co-authored-by: Michael Schlottke-Lakemper <michael@sloede.com>