Extend parallel P4estMesh mortar+AMR support to 3D (#1091)
* Use `p8est_load_ext` instead of `p8est_load`

`p8est_load` assumes that the number of processes is the same as when
the file was created. `p8est_load_ext` allows to load files with an
arbitrary number of processes.

* Extend parallel P4estMesh AMR/mortar support to 3D

* Update explanation of `P8EST_CONNECT_FULL` usage

* Clarify comments on container reinitialization during AMR

* Handle empty collections in MPI interface assert properly

Old version used generators which use `reduce` under the hood. Since
`reduce` throws an error on empty collections unless a neutral element
with respect to the operation in `reduce` is passed as an `init` value.
The the new version calls `mapreduce` explicitly so that such value can
be provided.

* Remove duplicate comment in `dg_3d_parallel.jl`

Co-authored-by: Michael Schlottke-Lakemper <michael@sloede.com>

* Add link to relevant p4est code

* Adjust coverrage override in p4est MPI test

Co-authored-by: Michael Schlottke-Lakemper <michael@sloede.com>