use PtrArray when wrapping with FDSBP to fix allocations
This has not been a problem when we created the FDSBP solvers. However, some updates in the package
ecosystem lead to spurious allocations that are hard to track down. A simple fix is to use PtrArrays
in Trixi.wrap_array. On Julia v1.8.5, this leads to a performance improvement of ca.
- 1.12s to 27.4ms for volume integral of examples/tree_1d_fdsbp/elixir_burgers_basic.jl
- 657ms to 112ms for volume integral of examples/tree_2d_fdsbp/elixir_advection_extended.jl
- 1.14s to 44.8ms for volume integral of examples/tree_3d_fdsbp/elixir_advection_extended.jl
With Julia v1.9, the previous situation was unbearable since there was a performance hit of roughly
a factor of 100.