reset threads also when initializing the summary callback (#1587)
* reset threads also when initializing the summary callback

I added the option to reset the threads from Polyester.jl in also in the summary callback.
The idea is that this supports another development workflow where we just modify the RHS
implementation and call solve again without re-ccreating the ODE.

The same comment as in 036eaed82b92be9376c5b610d8d40eddf45ca1fa applies:
However, I did not document it in the docstring since we have not documented
that we use Polyester.jl threads in general - and the resetting is specific
to Polyester.jl. I was not sure whether we still would like to keep the option
to change the threading backend any time - although I do not see a good reason
why we should do so.