Support `adapt` for SemiHyperbolic based on P4estMesh
This change adds support for `Adapt.adapt` for objects of type
`SemidiscretizationHyperbolic` based on `P4estMesh`. This allows
integration of such objects with KernelAbstractions.jl. In particular,
these objects can be moved to a GPU now. This is necessary to
facilitate

Some architectural changes and enhancements were necessary to achieve
this.

The type system around the container types was extended. The abstract
type `AbstractContainer` has an abstract subtype
`AbstractHeterogeneousContainer{T, B}` now that represents a
container that supports heterogeneous computing via KA.jl.
This allows information about the backend and whether KA.jl is used
to be propagated and used in downstream functions via the type system.
The type requires subtypes to implement a few methods, in particular
they must support Adapt.jl.

The p4est solver containers now subtype this new type and were changed
accordingly to support the required methods.

To add support for Adapt.jl, the macro Adapt.@adapt_structure
can be used in some cases to generate the required method.
However, the macro always uses the default constructor of a type
which uses the types of the arguments as type parameters. Most
types, in particular the container types, use other type
parameters and thus must use custom implementations of
the `Adapt.adapt_structure` function. In some cases, these
are currently not type stable (this is WIP).

Besides the container types, some solver-related types like
`LobattoLegendreBasis` and boundary condition types
also need to support Adapt.jl. This support was added as well.

As of now, equations usually use `SVector`s as parameters which
are bitstypes if their eltype is a bitstype and so they do not
have to be explicitly transfered to a GPU to be used in kernels.
Nonetheless, the equations are adapted when the
semidiscretization is adapted. By default this is a no-op
but allows for easier extension in the future if necessary.

The cache object is a `NamedTuple` and thus handled by a
standard method provided by Adapt.jl. Some parts of the
cache related to Mortars (e.g. the threaded caches) are
arrays of arrays where Adapt.jl will try to convert the
outer array by default which is not desired. Hence, a small
wrapper `VecOfArrays` is provided which implements
the desired behavior for Adapt.jl and is used for threaded
caches and other "vector of arrays" type objects that
need Adapt.jl support.

The SemidiscretizationHyperbolic constructors have been changed so that
the performance counter can be preserved when a semidiscretization is
adapted.