Implement basic 3D GPU kernels including MPI communication
Based on previously added support for Adapt.jl, it is now possible to
create an ODEProblem on a different backend by providing optional the
`adapt_to` keyword argument to `semidiscretize`. Under the hood
this calculates the initial condition as before and then calls
Adapt.jl on the semidiscretization and the initial solution.

The GPU kernels are mostly separated from existing code and are
reached via dispatching on the backend obtained from one of the
container types via the type system. The kernels are very basic
and do not include optimizations yet. The MPI communication works
but is very slow. In particular, a simulation with several GPUs may
be slower than a simulation with a single GPU.

The stepsize callback is explicitly GPU enabled while other callbacks
like the `AnalysisCallback` support the GPU by moving data to the CPU
and executing on the CPU afterwards.